realtek: build ZyXEL vendor firmware compatible initramfs
Append a version trailer used by the vendor firmware upgrade
application to validate firmwares.  The version names must
follow a strict naming scheme, and must evaluate to a version
higher than "V2.1".

The vendor firmware restricts the uload to the part of the
file described by the U-Boot Image header.  The version
trailer must therefore be inside the U-Boot Image.  And only
the initramfs image can be flashed from vendor firmware.

This allows OpenWrt to be installed from vendor firmware in
two steps:

 1) flash OpenWrt initramfs image from vendor web gui
 2) sysupgrade to a squasfs image once booted into OpenWrt

Note that OpenWrt can only be installed and run from the first
of the two system partitions.  The vendor firmware can be kept
in the second partition and be used as a backup. The next image
to boot can be selected from both vendor firmware and OpenWrt,
allowing true dual boot.

Signed-off-by: Bjørn Mork <bjorn@mork.no>