realtek: 6.6: fix VLAN handling
The CPU port of realtek switches usually accept a single VID. Because
the ethernet driver has no VLAN handling as of now we can only steer
packets by setting PVID for the CPU port. VLAN handling has never been
perfect but 3 events made things worse.

- Commit a376508 ("rtl83xx: dsa: Do nothing when vid 0")
- Commit e691e2b ("rtl83xx: dsa: reset PVID to 1 instead of 0")
- Upgrade to kernel 6.6

Reasons are:

- Rejecting VID 0 closed an open shortcut
- Initialization for PVID 1 forgot to set priv->ports[port].pvid
- Kernel 6.6 does no longer clarify CPU port as untagged

To fix this call the proper initialization routine during setup.
Ignore PVID settings for the CPU port for all further VLAN commands.

REMARK! The patch always binds VLAN 1 to the CPU port. We can see it
as a "management VLAN 1". So ensure that the management IP is bound
exactly to this VLAN.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>