realtek: fix untagged packets from CPU port
Commit  7c0949078d3e ("realtek: 6.6: fix VLAN handling") and the
two commits it fixes, makes VLAN 1 special. One consequence is that
ports with untagged VLANs must be member of VLAN 1 to avoid traffic
from the CPU port being filtered.

Fix by forcing all ports to be hidden members for VLAN 1.

This makes VLAN 1 even more magic. It is an acceptable compromise
which should not come as a surprise to most users. Switch firmware
from a number of vendors done similar things since the 1990's, and
VLAN 1 is often documented as a management only VLAN.

Signed-off-by: Bjørn Mork <bjorn@mork.no>